decoder: Refactor decoder generator names
- Add clear distinction between generator for
  regular vs. irregular instruction sets
