+++ vadl/main/vadl/vdt/passes/VdtLoweringPass.java
@@ -22,7 +22,7 @@ import vadl.configuration.GeneralConfiguration;
 import vadl.pass.Pass;
 import vadl.pass.PassName;
 import vadl.pass.PassResults;
-import vadl.vdt.impl.theiling.TheilingDecodeTreeGenerator;
+import vadl.vdt.impl.regular.RegularDecodeTreeGenerator;
 import vadl.vdt.model.Node;
 import vadl.vdt.utils.BitPattern;
 import vadl.vdt.utils.Instruction;
@@ -68,7 +68,7 @@ public class VdtLoweringPass extends Pass {
       return null;
     }
 
-    return new TheilingDecodeTreeGenerator().generate(insns);
+    return new RegularDecodeTreeGenerator().generate(insns);
   }
 
   /**
